[ptrnets] Add batch beam search decoding (**)
It's kinda bugged, because the decoding differs slightly when a
different batch size is used. I tested with the first 100 examples in
the test set, and for instance if I use a batch size of one compared to
two, there are three examples that are decoded differently. They
normally just differ in one or two  tokens, but still, it shouldn't be.